refactor code for get token logo
